lease enter a commit message to explain why this merge is necessary,
 Lines starting with '#' will be ignored, and an empty message aborts
i# the commit.

+++ app/Http/Controllers/MessageController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Requests\SendMessageRequest;
 use App\UserInfo;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\View;
@@ -13,7 +14,7 @@ use App\GoodCat;
 use App\GoodInfo;
 use App\Transaction;
 use App\TransactionLog;
-use App\Messages;
+use App\Message;
 use App\User;
 use App\Http\Controllers\Controller;
 use Storage;
@@ -24,44 +25,55 @@ class MessageController extends Controller
     {
         $data = [];
         $user_id = $request->session()->get('user_id');
-        $data['informations'] = Messages::Orderby('id','dsc')->where('receiver_id',$user_id)->get();
-        $data['users'] = User::orderBy('id', 'asc')->get()->keyBy('id');
-        return view::make('message.getMessage')->with($data);
+        $data['informations'] = Message::with('receiver')->Orderby('id','desc')->where('receiver_id',$user_id)->paginate(15);
+        return view::make('message.message')->with($data);
     }
 
-    public function sendMessagepage(Request $request)
+    public function getUnreadMsgNum(Request $request)
     {
-        return view::make('message.sendMessage');
-    }
-
-    public function editMessagepage(Request $request)
-    {
-        $data = [];
         $user_id = $request->session()->get('user_id');
-        $data['informations'] = Messages::Orderby('id','dsc')->where('receiver_id',$user_id)->get();
-        $data['users'] = User::orderBy('id', 'asc')->get()->keyBy('id');
-        return view::make('message.editMessage')->with($data);
+        $count = Message::where('receiver_id', $user_id)->where('is_read', false)->count();
+        return $count;
     }
 
-    public function deleteMessage(Request $request,$message_id)
+    public function readMessage(Request $request, $id)
     {
-        $data = Messages::find($message_id);
-        $data->delete();
-        return Redirect::to('/message/editmessage');
+        $data = Message::find($id);
+        if($request->session()->get('user_id') != $data->receiver_id)
+            return json_encode(['result' => false, 'msg' => 'Auth Failure.']);
+        $data->is_read = true;
+        $data->save();
+        return json_encode(['result' => true, 'msg' => 'success']);
     }
 
-    public function sendAllow(Request $request)
+    public function deleteMessage(Request $request, $id)
+    {
+        $data = Message::find($id);
+        if($request->session()->get('user_id') != $data->receiver_id)
+            return json_encode(['result' => false, 'msg' => 'Auth Failure.']);
+        $data->delete();
+        return json_encode(['result' => true, 'msg' => 'success']);
+    }
+
+    public function sendMessage(SendMessageRequest $request)
     {
         $input = $request->all();
-        $message = new Messages;
+        if(!User::where('id', $input['receiver'])->count())
+            return json_encode(['result' => false, 'msg' => 'no such receiver']);
+        $message = new Message;
         $user_id = $request->session()->get('user_id');
         $message->title = $input['title'];
         $message->content = $input['content'];
         $receiver = $input['receiver'];
         $message->sender_id = $user_id;
-        $user = User::where('nickname',$receiver)->first();
+        $user = User::where('id', $receiver)->first();
         $message->receiver_id = $user->id;
         $message->save();
-        return Redirect::to('/message');
+        return json_encode(['result' => true, 'msg' => 'success']);
+    }
+
+    public function sendMessagepage(Request $request)
+    {
+        return view::make('message.sendMessage');
     }
 }
